auto selector works

+++ 2020-Robot/src/main/java/frc/team670/robot/commands/auton/AutoSelector.java
@@ -81,7 +81,7 @@ public class AutoSelector {
             ID = id;
         }
 
-        public int getID() {
+        public double getID() {
             return this.ID;
         }
